+++ example/get_list.dart
@@ -1,5 +1,7 @@
 import 'package:ps_list/ps_list.dart';
 
 Future<void> main(List<String> args) async {
-  PSList.getRunningProcesses();
+  final ps = await PSList.getRunningProcesses();
+
+  print(ps);
 }